+++ packages/dialtone-vue2/components/emoji_picker/emoji_picker.test.js
@@ -48,6 +48,24 @@ const MOCK_RECENTLY_USED_EMOJIS = [
     unicode_character: '1f470-1f3ff-2640',
   },
 ];
+const MOCK_CUSTOM_EMOJIS = [
+  {
+    name: 'shipit',
+    date_added: 1730918816847,
+    added_by: 'Ignacio Ropolo',
+    image: 'https://github.githubassets.com/images/icons/emoji/shipit.png',
+    unicode_character: '1f44d',
+  },
+  {
+    name: 'thumbs up',
+    category: 'people',
+    shortname: ':thumbsup:',
+    shortname_alternates: [':+1:', ':thumbup:'],
+    keywords: ['+1', 'hand', 'thumb', 'up', 'uc6'],
+    unicode_output: '1f44d',
+    unicode_character: '1f44d',
+  },
+];
 const MOCK_TAB_SET_LABELS = [
   'Most recently used',
   'Smileys and people',
@@ -58,6 +76,7 @@ const MOCK_TAB_SET_LABELS = [
   'Objects',
   'Symbols',
   'Flags',
+  'Custom',
 ];
 const MOCK_SKIN_SELECTOR_BUTTON_TOOLTIP_LABEL = 'Change default skin tone';
 const MOCK_SEARCH_RESULTS_LABEL = 'Search results';
@@ -66,6 +85,7 @@ const MOCK_SEARCH_PLACEHOLDER_LABEL = 'Search...';
 const baseProps = {
   skinTone: 'Light',
   recentlyUsedEmojis: MOCK_RECENTLY_USED_EMOJIS,
+  customEmojis: MOCK_CUSTOM_EMOJIS,
 };
 
 let mockProps = {};
@@ -130,6 +150,13 @@ describe('DtEmojiPicker Tests', () => {
       expect(firstButton.attributes('aria-label')).toBe(MOCK_TAB_SET_LABELS[0]);
     });
 
+    it('Should render customs emojis tabset', () => {
+      const lastButton = wrapper.find('.d-tablist').findAll('button').at(MOCK_TAB_SET_LABELS.length - 1);
+
+      expect(lastButton.exists()).toBe(true);
+      expect(lastButton.attributes('aria-label')).toBe(MOCK_TAB_SET_LABELS[MOCK_TAB_SET_LABELS.length - 1]);
+    });
+
     it('Should render provided search placeholder label', () => {
       const searchInput = wrapper.find('.d-emoji-picker__search input');
 
@@ -148,6 +175,13 @@ describe('DtEmojiPicker Tests', () => {
       const TabsCount = wrapper.find('.d-tablist').findAll('button').length;
 
       expect(TabsCount).toBe(MOCK_TAB_SET_LABELS.length);
+    });
+
+    it('Should render add emoji button', () => {
+      const addEmojiButton = wrapper.find('.d-emoji-picker__add-emoji');
+
+      expect(addEmojiButton.exists()).toBe(true);
+      expect(addEmojiButton.attributes('aria-label')).toBe('Add emoji');
     });
 
     describe('Skin tone selector tests', () => {
@@ -323,10 +357,10 @@ describe('DtEmojiPicker Tests', () => {
     });
 
     it('Should jump to skin selector from emoji-selector', async () => {
-      const firstFlags = wrapper.find('.d-emoji-picker__selector .d-emoji-picker__alignment:nth-child(10) button');
+      const firstCustoms = wrapper.find('.d-emoji-picker__selector .d-emoji-picker__alignment:nth-last-child(1) button');
       const skinSelector = wrapper.find('.d-emoji-picker__skin-selected button');
 
-      await firstFlags.trigger('keydown.tab');
+      await firstCustoms.trigger('keydown.tab');
 
       expect(document.activeElement).toBe(skinSelector.element);
     });