+++ packages/dialtone-vue2/components/emoji_picker/emoji_picker_default.story.vue
@@ -2,11 +2,14 @@
   <dt-emoji-picker
     :skin-tone.sync="isSkinTone"
     :recently-used-emojis="$attrs.recentlyUsedEmojis"
+    :custom-emojis="$attrs.customEmojis"
     :search-query="$attrs.searchQuery"
     :show-search="$attrs.showSearch"
     @skin-tone="isSkinTone = $event; $attrs.onSkinTone($event)"
     @close="$attrs.onClose($event)"
     @selected-emoji="$attrs.onSelectedEmoji($event)"
+    @scroll-bottom-reached="$attrs.onScrollBottomReached($event)"
+    @add-emoji="$attrs.onAddEmoji($event)"
   />
 </template>