+++ packages/dialtone-vue2/components/rich_text_editor/rich_text_editor.test.js
@@ -89,7 +89,7 @@ describe('DtRichTextEditor tests', () => {
             content: [{
               type: 'paragraph',
               attrs: {
-                textAlign: 'left',
+                textAlign: null,
               },
               content: [{
                 text: 'new value',