+++ packages/dialtone-vue2/components/rich_text_editor/rich_text_editor.vue
@@ -599,7 +599,6 @@ export default {
 
       extensions.push(TextAlign.configure({
         types: ['paragraph'],
-        defaultAlignment: 'left',
       }));
 
       if (this.allowCode) {