+++ packages/dialtone-vue2/recipes/conversation_view/message_input/message_input_default.story.vue
@@ -44,6 +44,7 @@
         @json-input="$attrs.onJsonInput"
         @html-input="$attrs.onHtmlInput"
         @text-input="$attrs.onTextInput"
+        @add-emoji="$attrs.onAddEmoji"
         @select-media="$attrs.onSelectMedia"
         @selected-emoji="$attrs.onSelectedEmoji"
         @selected-command="$attrs.onSelectedCommand"