+++ packages/dialtone-vue3/components/rich_text_editor/rich_text_editor.vue
@@ -597,7 +597,6 @@ export default {
 
       extensions.push(TextAlign.configure({
         types: ['paragraph'],
-        defaultAlignment: 'left',
       }));
 
       if (this.allowCode) {